// src/redux/actions/actions.types.ts
export enum ACTION_TYPES {
  GET_CHARACTERS = "GET_CHARACTERS",
  SEARCH_CHARACTERS = "SEARCH_CHARACTERS",
  STATUS_SEARCH = "STATUS_SEARCH",
  NAME_SEARCH = "NAME_SEARCH",
  SPECIES = "SPECIES",
  CHECKBOX = "CHECKBOX",
  CHECK_ALL_CHECKBOX = "CHECK_ALL_CHECKBOX",
  REMOVE_CHARACTERS = "REMOVE_CHARACTERS",
  CHANGE_STATUS = "CHANGE_STATUS",
  CHANGE_PAGE = "CHANGE_PAGE",
  ADD_RESULTS = "ADD_RESULTS",
}